ci: automate release draft
When the version-bump PR gets merged, the release draft workflow starts
automatically. This is indicated by a comment on the PR.

The release draft workflow creates the source tarball (which includes
all subprojects), builds Criterion from the tarball, and creates binary
packages.

Binary packages are created within a CentOS 7 container. This sounds
terrible, but it ensures that Criterion is built with an older glibc
version (v2.17) so that it will work on a wide variety of Linux distros.

The source tarball and the binaries then will be attached to a release
draft on the Releases page.